policy/requirements.txt
psycopg2-binary
kubernetes
httpx
